+++ lib/acts_as_railable.rb
@@ -1,6 +1,8 @@
 require "acts_as_railable/version"
 require "acts_as_railable/railtie"
 require "acts_as_railable/core_ext"
+require "acts_as_railable/interval"
+require "acts_as_railable/humanable"
 
 module ActsAsRailable
   # Your code goes here...