+++ test/dummy/app/models/application_record.rb
@@ -1,3 +1,6 @@
 class ApplicationRecord < ActiveRecord::Base
+  include ActsAsRailable::Humanable
+  include ActsAsRailable::Interval
+
   primary_abstract_class
 end