+++ CHANGELOG.md
@@ -4,12 +4,17 @@
 ## Unreleased
 
 ### Features
-* Interval Module
-* Humanable Module
 * Recordable Module
 
 ### Bugfixes
 * 无
+
+
+## v0.2.0 (2022-04-01)
+
+### Features
+* Interval Module
+* Humanable Module
 
 
 ## v0.1.0 (2022-04-01)